OneDrive/Desktop/MusicPlayer with python/requirements.txt
pip==21.0.1
python-vlc==3.0.12118
setuptools==56.0.0
